% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2015HarveyLiu.R
\name{BonfAdj}
\alias{BonfAdj}
\title{Bonferroni adjustment}
\usage{
BonfAdj(pvals)
}
\arguments{
\item{pvals}{A vector of p-values.}
}
\value{
The Bonferroni-adjusted vector of p-values.
}
\description{
Adjusts a vector of p-values based on Bonferroni's method.
}
\examples{
data(IS)

TT <- ncol(IS[, 2:6])
SR_hat <- apply(IS[, 2:6], 2, SharpeRatio)
t_ratios <- SR_hat * sqrt(TT)
plot(sort(t_ratios), type = "l")

pvals <- 1 - unlist(lapply(t_ratios, pt, df = TT - 1))
pvals

BonfAdj(pvals)
}
