% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{IS}
\alias{IS}
\title{In-sample strategy returns}
\format{
An object of class \code{data.frame} with 1258 rows and 101 columns.
}
\usage{
data(IS)
}
\description{
In-sample returns of 100 momentum strategies applied to the SP500 using different parameter combinations.
}
\examples{
data(IS)
apply(IS, 2, SharpeRatio, rfr = 0.07)
}
\keyword{datasets}
