% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2014BaileyEtAl.R
\name{MinBTL}
\alias{MinBTL}
\title{Minimum Backtest Length}
\usage{
MinBTL(m = NULL, rfr = 0, annualize_by = 1, na.rm = FALSE, N, SR_hat)
}
\arguments{
\item{m}{A matrix of returns.}

\item{rfr}{The risk-free (rate of) return. Only used if m != NULL.}

\item{annualize_by}{The factor used to annualized returns. Only used if m != NULL.}

\item{na.rm}{logical. Should missing values be removed when estimating the Sharpe ratios? Only used if m != NULL.}

\item{N}{The number of strategies/parameter combinations tested. Only used if m = NULL.}

\item{SR_hat}{The IS Sharpe ratio for which the MinBTL should be estimated. Only used if m = NULL.}
}
\value{
The Minimum Backtest Length (MinBTL) and its upper bound (UpperBound).
}
\description{
Estimates the Minimum Backtest Length (MinBTL) and its upper bound for a matrix of strategy returns or a number of strategies and a IS Sharpe ratio.
If a matrix (m) is being provided, the function uses its column number as the number of strategies and the maximum Sharpe ratio across its columns as the IS Sharpe ratio. If no matrix is being provided (m = NULL), the arguments N and SR_hat are being used instead.
}
\examples{
data(IS)
MinBTL(IS[, 2:6], rfr = 0, annualize_by = 1, na.rm = FALSE)
MinBTL(m = NULL, n = 10, SR_hat = 1.5)
}
