% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2012BaileyLopez.R
\name{MinTRL}
\alias{MinTRL}
\title{Minimum Track Record Length}
\usage{
MinTRL(
  r,
  rfr = 0,
  SR_bench = 0,
  alpha = 0.05,
  annualize_by = 1,
  round_up = TRUE,
  na.rm = FALSE
)
}
\arguments{
\item{r}{A vector of returns.}

\item{rfr}{The risk-free (rate of) return.}

\item{SR_bench}{A benchmark Sharpe ratio}

\item{alpha}{The significance level.}

\item{annualize_by}{The factor used to annualized returns.}

\item{round_up}{Whether the resulting MinTRL should be rounded up or not.}

\item{na.rm}{logical. Should missing values be removed?}
}
\value{
The (annualized) Sharpe ratio.
}
\description{
Computes the Minimum Track Record Length (MinTRL) for a vector of returns.
}
\examples{
data(IS)
MinTRL(IS[, 2], 0, 0, 0.05, 1)
}
