% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{OOS}
\alias{OOS}
\title{Out-of-sample strategy returns}
\format{
An object of class \code{data.frame} with 1258 rows and 101 columns.
}
\usage{
data(OOS)
}
\description{
Out-of-sample returns of 100 momentum strategies applied to the SP500 dataset using different parameter combinations.
}
\examples{
data(OOS)
apply(OOS, 2, SharpeRatio, rfr = 0.07)
}
\keyword{datasets}
