% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2012BaileyLopez.R
\name{PSR}
\alias{PSR}
\title{Probabilistic Sharpe ratio}
\usage{
PSR(r, rfr = 0, SR_bench = 0, annualize_by = 1, na.rm = TRUE)
}
\arguments{
\item{r}{A vector of returns.}

\item{rfr}{The risk-free rate of return.}

\item{SR_bench}{A benchmark Sharpe ratio}

\item{annualize_by}{The factor used to annualized returns.}

\item{na.rm}{logical. Should missing values be removed?}
}
\value{
The (annualized) Sharpe ratio.
}
\description{
Computes the probabilistic Sharpe ratio (PSR) for a vector of returns.
}
\examples{
data(STRATEGIES)
PSR(STRATEGIES[, 2], 0, 0, 1)
}
