% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2012BaileyLopez.R
\name{SEF}
\alias{SEF}
\title{Sharpe Ratio Efficient Frontier}
\usage{
SEF(m, rfr = 0, annualize_by = 1, na.rm = FALSE)
}
\arguments{
\item{m}{A matrix of strategy returns.}

\item{rfr}{The risk-free rate of return.}

\item{annualize_by}{The factor used to annualized returns.}

\item{na.rm}{logical. Should missing values be removed?}
}
\value{
The (annualized) Sharpe ratio.
}
\description{
Computes the Sharpe Ratio Efficient Frontier (SEF) for a vector of returns.
}
\examples{
data(STRATEGIES)
SEF(STRATEGIES[,-1], 0, 1)
}
