% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basic_functions.R
\name{SharpeRatio}
\alias{SharpeRatio}
\title{Sharpe ratio}
\usage{
SharpeRatio(r, rfr = 0, annualize_by = 1, na.rm = FALSE)
}
\arguments{
\item{r}{A vector of returns.}

\item{rfr}{The risk-free rate of return.}

\item{annualize_by}{The factor used to annualized returns.}

\item{na.rm}{logical. Should missing values be removed?}
}
\value{
The (annualized) Sharpe ratio.
}
\description{
Computes the Sharpe ratio (SR) for a vector of returns.
}
\examples{
SharpeRatio(rnorm(252), 0, 1)
}
