% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{SP500}
\alias{SP500}
\title{S&P 500 index price data}
\format{
An object of class \code{data.frame} with 2776 rows and 6 columns.
}
\usage{
data(SP500)
}
\description{
Price data of the S&P 500 index between the 2012-12-19 and 2023-12-31.
}
\examples{
data(SP500)
plot(SP500$Adjusted, type = "l")
}
\keyword{datasets}
