% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{STRATEGIES}
\alias{STRATEGIES}
\title{Strategy returns}
\format{
An object of class \code{data.frame} with 2516 rows and 101 columns.
}
\usage{
data(STRATEGIES)
}
\description{
Returns of 100 momentum strategies applied to the SP500 using different parameter combinations.
}
\examples{
data(STRATEGIES)
apply(STRATEGIES, 2, SharpeRatio, rfr = 0.07)
}
\keyword{datasets}
